refactor(cli): modularize serve command and improve CLI structure
- Extract serve command logic into separate file
- Create index.ts for serve command group
- Implement serve.ts with detailed server setup
- Update main entry point to use new command structure
- Refactor logger configuration for global use
